Change: item attribute from float to double

+++ src/main/java/me/adamix/mercury/server/item/attribute/ItemAttributeValue.java
@@ -2,5 +2,5 @@ package me.adamix.mercury.server.item.attribute;
 
 import net.minestom.server.entity.attribute.AttributeOperation;
 
-public record ItemAttributeValue(float value, AttributeOperation operation) {
+public record ItemAttributeValue(double value, AttributeOperation operation) {
 }
